requirements.txt
easyocr
paddleocr
paddlepaddle
google-generativeai
